Made a bunch of changes to the VI in order to make the timing and such run better
Added a rendezvous to the timing for the photodiodes and some of the
pressure and temperature sensor readings

Changed the way the predet runs so that there is not state machine and
it is just a sequence that is controlled by the error flow. This does
not allow for stopping in the middle of the run process currently but
that has not been an issue so far.

Removed a bunch of subVIs and replaced them with other subVIs that
contain only the setup command blocks. This was done in an attempt to
make the timing easier so that the execution or running of each task
occurs in the main vi